+++ ios/UseHealthKit/Query.swift
@@ -1,0 +1,40 @@
+import Foundation
+import HealthKit
+
+class Query {
+    /// Get sample data of BodyMass.
+    /// - Parameters:
+    ///   - type: type to get data
+    ///   - options: <#options description#>
+    ///   - startDate: start date to get data.
+    ///   - endDate: end date to get data.
+    ///   - predicate: <#predicate description#>
+    ///   - anchorDate: <#anchorDate description#>
+    ///   - interval: interval of collection. Default value is 1 day.
+    ///   - completion: handler when the query completes.
+    static func makeHKStatisticsCollectionQuery(_ type: HKQuantityType,
+                                                _ options: HKStatisticsOptions,
+                                                _ startDate: Double,
+                                                _ endDate: Double,
+                                                _ predicate: NSPredicate? = nil,
+                                                _ anchorDate: Date? = nil,
+                                                _ interval: DateComponents? = nil) -> HKStatisticsCollectionQuery {
+        let _predicate = HKQuery.predicateForSamples(withStart: Date(timeIntervalSince1970: startDate),
+                                                     end: Date(timeIntervalSince1970: endDate),
+                                                     options: [.strictStartDate, .strictEndDate])
+
+        let from = Date(timeIntervalSince1970: startDate)
+        var anchorDateComponent = Calendar.current.dateComponents([.year, .month, .day], from: from)
+        anchorDateComponent.hour = 0
+        let _anchorDate = Calendar.current.date(from: anchorDateComponent)!
+
+        var _interval = DateComponents()
+        _interval.day = 1
+
+        return HKStatisticsCollectionQuery(quantityType: type,
+                                           quantitySamplePredicate: (predicate != nil) ? predicate! : _predicate,
+                                           options: options,
+                                           anchorDate: (anchorDate != nil) ? anchorDate! : _anchorDate,
+                                           intervalComponents: (interval != nil) ? interval! : _interval)
+    }
+}
